+++ user-service/src/main/java/cuit9622/dms/user/service/impl/MenuServiceImpl.java
@@ -13,11 +13,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-/**
- * @author lsh
- * @description 针对表【sys_menu】的数据库操作Service实现
- * @createDate 2024-04-13 16:56:49
- */
+
 @Service
 public class MenuServiceImpl extends ServiceImpl<MenuMapper, Menu>
         implements MenuService {
@@ -26,9 +22,9 @@ public class MenuServiceImpl extends ServiceImpl<MenuMapper, Menu>
     MenuMapper menuMapper;
 
     @Override
-    public List<MenuTree> getTreeMenu(Integer userId) {
+    public List<MenuTree> getTreeMenu(Long userId) {
         List<Menu> menus = menuMapper.getMenuByUserId(userId);
-        Map<Integer, MenuTree> map = new HashMap<>();
+        Map<Long, MenuTree> map = new HashMap<>();
         List<MenuTree> list = new ArrayList<>();
         // 将数据全部放在map集合中
         for (Menu menu : menus) {
@@ -43,11 +39,11 @@ public class MenuServiceImpl extends ServiceImpl<MenuMapper, Menu>
                 map.put(menu.getMenuId(), menuTree);
             }
         }
-        for (Map.Entry<Integer, MenuTree> entry : map.entrySet()) {
+        for (Map.Entry<Long, MenuTree> entry : map.entrySet()) {
             MenuTree value = entry.getValue();
             List<MenuTree> children;
             // 以父id为key如果能查到代表该节点有子节点
-            Integer parentId = value.getMenu().getParentId();
+            Long parentId = value.getMenu().getParentId();
             MenuTree parent = map.get(parentId);
             if (parent != null) {
                 children = parent.getChildren();
@@ -61,7 +57,3 @@ public class MenuServiceImpl extends ServiceImpl<MenuMapper, Menu>
         return list;
     }
 }
-
-
-
-
